Refactor Pin class
