--- app/src/simulator/core/Pin.kt
@@ -1,8 +1,0 @@
-package simulator.core
-
-import java.awt.Point
-
-class Pin(val name: String,
-          var value: Int?,
-          val pos: Point,
-          val relativeElement: Any)
